% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.func.1D.R
\name{fit.func.1D}
\alias{fit.func.1D}
\title{Estimate the functions and functional derivatives via B-spline}
\usage{
fit.func.1D(Y, X, X.pred = NULL, nknots = NULL, d = 3, derivs = 1,
  X.sup = NULL, GCV.flag = TRUE)
}
\arguments{
\item{Y}{a matrix of one-dimensional dense functional data with each row corresponding to one subject.}

\item{X}{an optional vector containing locations of observations.
If not specified, the data are treated to be observed on regular grids in \code{(0,1])}.}

\item{X.pred}{an optional vector specifying the locations on which the estimated mean or derivative
functions are returned in the output. If not provided, estimations will returned on the same locations as in \code{X}.}

\item{nknots}{a vector or an integer specifying the potential numbers of knots used in B-spline estimation.
Default is \code{seq(2,min(floor(n/4),20),1)}, where \code{n} is the total number of subjects.}

\item{d}{degree of  B-spline for estimating mean function, default is 3 for cubic splines.}

\item{derivs}{order of functional derivatives to estimate. Default is 1.}

\item{X.sup}{a length 2 vector of the boundary knots. Default is minimum and maximum in \code{c(X, X.pred)}.}

\item{GCV.flag}{a logical value indicating whether generalized cross validation should be used to select \code{nknots}.}
}
\description{
The function is used to estimate mean function of one-dimensional functional data and its derivatives.
The number of knots is selected using generalized cross validation.
}
