% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.func.R
\name{plot.func}
\alias{plot.func}
\title{Plot estimation and sccs of mean function and its derivatives for functional data}
\usage{
\method{plot}{func}(mfit)
}
\arguments{
\item{mfit}{an "func" object returned from either \code{fit.func.1D} or \code{scc.1D}.}
}
\description{
The plot method for class 'func'. The function plots estimated mean functions (derivatives) and corresponding simultaneous confidence corridors.
}
